Moved and updated current resources
