+++ graphics/resources/Animation.cs
