+++ graphics/resources/Material.cs
